refactor(src): rename index.ts to main.ts and put exports in index.ts
